requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
pyyaml>=6.0
apscheduler>=3.10.4
mcp>=0.9.0
